pyproject.toml
[project]
name = "mlops-labs"
version = "0.1.0"
description = "MLOps Labs: Chest X-Ray Pneumonia Classification"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "rootutils",
    "pandas",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "mlflow",
    "pillow",
    "tqdm",
    "kagglehub",
    "typer",
    "optuna>=4.7.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
]

[dependency-groups]
dev = [
    "ipykernel",
    "mypy",
    "pre-commit",
    "ipywidgets",
    "ruff",
    "pytest>=9.0.2",
    "apache-airflow>=3.1.7",
]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
testpaths = ["tests"]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",    # line too long, handled by black
    "B008",    # do not perform function calls in argument defaults
    "C901",    # too complex
    "D1",      # Allow missing docstrings.
    "D417",    # Don't require descriptions for all arguments.
    "F401",    # unused imports
]
[tool.mypy]
ignore_missing_imports = true

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
force-sort-within-sections = true
combine-as-imports = true
split-on-trailing-comma = false
lines-after-imports = 2

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
multiline-quotes = "double"
inline-quotes = "double"
